Adjust tests to use libgpr2 instead of the tools
gpr-symlinks used gprinstall and gprclean to check proper handling of
symlinks in libgpr2, which made the tests N/A for the next branch for which
those tools don't exist yet.

Changing the tests to directly use libgpr2 allows to test both branches.

Part of the regular merge master -> next

no-issue-check