Fix elaboration issue with config project requiring externals
If a config project requires an external value, and this value is defined
by the root aggregate project, then it was not taken into account while
loading the config, leading to an error.

This is now fixed by evaluating the root project before the configuration
project in case it is an aggregate.