Add a customizable log message reporter
Now that Tree.Load function displays automatically the messages, we need to
allow users to customize this output, in particular in case stdout/stderr are
not available for that.

This change now provides in GPR2.Project.Tree an homogeneous way to handle
messages from the tree or from updating sources.