Update the provided action in `Add_Action`
A copy of the provided action is inserted in the tree_db actions. This copy
is then modified by subprograms like Attach. If the user wants to add an
action and keep using it afterwards, then this action will not be sync
with the one in the tree DB. In particular, its `Tree` field will still
be null, despite the action being added.

This field is used, for instance, by the `Ada_Compile` action. If a user
creates their own action and manipulates it without querying the tree db after
a call to `Add_Action`, then the action state would be incorrect.