Compute signatures from GPR2.Build.Actions
The method of comparing signatures has been updated. Previously, all inputs
and outputs registered were searched in the JSON and compared afterward.
This approach is flawed because an artifact in the saved signature may no
longer exist in the code, thus avoiding verification.

The current approach reads the entire stored signature and compares its
checksum with the checksums of the artifacts stored internally.

This process is applied universally to all actions, so the function
Compare_Signature is no longer abstract.