Fix view closure computation when an aggregate library is present
View closure is computed during the first parse stage, but aggregated
projects are computed later when the Project_Files attribute is available.

As a result, aggregate libraries view closure is incomplete and not propagated
to including views. A special handling thus needs to be done.

In the context of eng/gpr/gpr-issues#288