Allow abstract project importing non-abstract projects
When updating sources, each source has its visibility updated. For example,
if a source belongs to an extended view, the source will have its unit in the
extending view rather than the extended one.

This visibility check on abstract projects raised a constraint exception
because it required access to the view database, which is not permitted for
abstract views.

Since abstract views do not have sources by definition, this will not impact
the visibility check.