Adjust the way the runtime project is handled
This change ensures that only projects actually having Ada sources import
the runtime project.

This also changes this import from regular with to limited with import,
the rationale being that the runtime project itself contains no actual
useful attribute to the user projects, while it may clash with existing
use projects named "runtime". This allows us to add the view late in the
parsing project, so that attributes can be evaluated to understand if the
views are concrete and define the Ada language.

Preparatory work for eng/gpr/gpr-issues#332