Adjust the Ada_Compile action
Store the involved compilation unit in the action, to ease its lookup.
Improve the display of the action.
Make the UID type of the action a constrained type.
Fix the lookup of source dependencies in Update_Deps_From_Ali.
Ada_Compile.Post_Bind: do not override subprograms that don't need to be.