Rework how the initial DAG is generated
In particular this introduces a new On_Tree_Propagation primitive in
the Actions to allow complete the DAG by the actions themselves.

Use this mechanism to have the Ada_Compile action properly populate the
Bind and Link phases based on dependencies from ALI or from the Ada
parser.

The initial DAG population now handles both libraries and executables.