Improve signature load/store and check
This change introduces a way to serialize/deserialize artifacts with
an uri-based encoding, allowing the signature to store and load artifact
objects directly.

The signature is then simplified to just contain a map of Artifacts with
their corresponding checksum. The stored json file is also simplified.

This removes the temporary states "Valid" and "Coherant", source of bugs in
general since in practice the actual artifact can change any time, so saving
these states is dangerous.

The signature checking explicit stage is thus removed, and is done upon
call to the "Valid" function. As this is done only once before executing
and action it's fine to not try to optimize it here.