gprinstall: Support the --autoconf option
The --autoconf option specifies a configuration project file name that already
exists or will be created automatically. Option --autoconf= cannot be specified
more than once. If the configuration project file specified with --autoconf=
exists, then it is used. Otherwise, GPRconfig is invoked to create it
automatically.

This option was removed during the port of gprinstall to libgpr2.