Fix assertion error in gprname
In Check_Name_Validity, make sure that we we don't have a Sloc, then
no Assertion_Error is raised. This means in particular we can't log the
message in this case but that was useless anyway since just the status
of the messages was check. Return a Boolean instead.

In View_Internal, don't raise a fatal error when there's a duplicated entry
in non-ada naming exception, since this case is overall harmless. Change that
to a warning instead.

In the context of eng/gpr/gpr-issues#147