Adapt tools to the new reporter system and simplify the tools verbosity
The GPRtools option object now includes an internal console reporter, with its
properties affected by the command line options. This reporter is then passed to
the Load function to be used by the project tree. For tools that do not require
loading a tree, such as gprconfig, a local reporter is created.

When the -ws flag is used to disable warnings, the reporter's verbosity is set
to No_Warnings. This means that only error messages will be shown in this mode,
unlike the previous implementation where lint, informational, and warning
messages could be processed independently.