Replace the GPR2.Message Info level with Hint and introduce the End_User level.
The End_User level should be used for messages that do not require formatting.
For example, it can be used by tools to indicate which project file is being
used or that a directory has been created.

The Reporter's interface now includes only one Internal_Report abstract method,
as simple strings are replaced by messages with the End_User level.

Additionally, it is now possible to create messages that do not have a source
location reference. These messages will be prefixed only with their level value.