Adjust types handling in the project parser
The parsing of the string types declaration was done in stage 1
of the parsing while it is not needed there. It is now done together
with the other expressions in the main parsing Procedure.

This allows for a better display of errors and warnings when such
error involves types declaration: previously only this specific
error was displayed, the general case being then skipped.