Adjust the message levels when loading sources
This change use a Warning level when there's clash of non-compilable
sources instead of the Error level.

This also introduces a new option in Tree.Update_Sources to allow
preventing errors to be generated, useful for applications that
don't fail upon invalid source set (such as an IDE).

Fix for eng/gpr/gpr-issues#401