Refactoring of External_Options
Move External_Options to the build directory and remove all mentions
of it from the lib hierarchy (in particular GPR2.Project.Tree but
also GPR2.Options): this is only used by the build database.

Also simplify the Tree_Db API so that specifying it is now optional
and asynchronous to the load of the tree.

Add the option to GPRbuild.Options, since this is the only tool
for which this is useful.