Strengthen the command line computation phase
In case we need to create a temp file and the object directory is
missing/not writable/not executable we need to gracefully handle
the OS_Error exception and report the issue to the end user.

Also don't attempt to compute the command line if the corresponding
view is externally built or the action is deactivated.