Ada_Bind action: rework the support for standalone libraries
This removes the generic Extra_Argument when creating a bind
action, moving the knowledge of which switch to use in case of
standalone libraries from action_population to the ada bind action
itself.

This adds SAL_In_Closure parameter and Has_Main parameter.

Add proper means in actions_population to track the standalone
libraries in the sub-closure of a view.