Compile Ada action: take care of standalone libraries
For standalone libraries, only the ALI files of interface sources
are to be copied in the library ali directory (and amended to
contain the SL flag). This means that lookup of inherited ali file
needs to be amended to take care of ali files only present in the
object directory.