Add the linkage runtime flags for non-standalone shared libraries
When linking an executable or a standalone library, a binding phase is included.
This binding phase provides the necessary flags to the linker.
However, this binding phase does not occur when linking a non-standalone shared
library. Consequently, the -l and -L options must be added manually.

To avoid runtime conflicts, the exact version of the runtime must be determined.
This is achieved by parsing the system ALI file.