Improve support for aggregate and standalone libraries.
This change better controls the various version of an ali file
after it's produced by the compile action: the various libraries
it's copied into and with SL flag or untouched.

The Bind action is refined to properly use the local ali files
located in the obj directory, and the lib variants for library
imports.

For eng/gpr/gpr-issues#474