Modify the condition to use the linker options provided by the binder, if any
Previously, only executables benefited from the linker options provided by
the binder. This caused issues for partially linked static standalone libraries
which were not linked with the runtime. This patch ensure that the following
entities benefit from these options:
* executables
* shared libraries
* static standalone libraries that are partially linked.

Note that the last point is not covered yet, as we do not support
partial linking.