Bind action: gnatbind prefixes and path may come from all switches attributes
This MR makes sure that both Required_Switches and Switches may define
the arguments allowing to pick up the right gnatbind.

As a result the gnatbind driver can only be set after most of the
command line has been processed: so a new Set_Driver primitive is
required in the command line object.

Adjust the other actions to use this new primitive.

This change also took the opportunity to replace the use of string
allocation with Unbounded_String or Path_Name objects.

Issue found in the context of eng/gpr/gpr-issues#474