Adjust retrieval of sources by basename
Having several sources with the same basename and may be
problematic when wanting to retrieve two sources that are ambiguous
since they belong to two distinct projects that are just withed.

This change first detects such case and reports the ambiguity. It
also provides a new mean to retrieve a source from its full path.

In the context of eng/gpr/gpr-issues#474