Rework the artifacts API
This removes the Source artifact, that was more trouble then
useful.

This introduces a new Key_Value artifact allowing to store
command line and compiler version in signatures.

This changes thus the way the artifacts are serialized/unserialized
to allow this key_value support.

This also removes the SLOC primitive, only applicable to files.

This adds a View item, allowing to reference the view of an
artifact if needed