Rework the action's signatures
This changes the way the artifacts are saved/restored and checked

The format for artifact is now:
{"class": "",
 "key": "",
 "value": ""}

The class allowing to find a proper instance, then then
unserializes itself using key and value.

The value is generic: for a file it will be a checksum but for
key-value it will be the actual value, where all is needed is
a simple string comparison.

When loading the signature, the actions now populate the
expected artifacts manually before run. Each artifact is checked
against the signature, and this process stops as soon as there
is a mismatch. This allows for reliable and efficient checks.

On the other hand, when saving the signature, the final list of
artifacts is populated and their checksums/values dumped.