Add the link options from the binder only after the library dependencies
The link options provided by the binder typically include the runtime link
option. When linking an executable with a static library that depends on
the runtime, the linker (ld) expects the library archive to be specified before
the runtime archive in the link command line. This ensures that symbols are
correctly resolved.

Tests have been improved to detect this type of errors.