Fix pragma files signature in compile.ada without -gnateb
This is still work-in-progress, but local and global pragma files
cannot be treated as regular sources, since they're not considered
as such by the build system. Further work will be needed to account
for -gnateb that removes the pre-condition of this test: pragma files
from the ali file should be full paths.

For eng/gpr/gpr-issues#474