Allow checking signatures without creating temporary files.
The command line computation needs to be performed for signature
checking, so this MR introduces a mode where no temporary file is
created in this context, but where the command line is thus
incomplete.

This is needed to allow signature checking for tools that should not
create any artifact.

For eng/gpr/gpr-issues#474