Adjust the computation of Standalone Library Interfaces
This change ensures that the ali files of the completed interface
are copied to the library directory. So including the units that
are implicitly in the interface closure.

To do so the copy phase has to be done at link time where the
complete interface is known (it cannot be done at bind time since
non standalone libraries don't have an associated bind phase).

A new dependency must then be added between the bind actions and
their dependency library link actions, so that bind occurs only
after the complete set of ali files are exported to the library
directories.

For eng/gpr/gpr-issues#474