Implement the support of linker options for static standalone libraries
When a static library is created, the linker options required to link
with it are stored in a dedicated section .GPR.linker_options. These options
can later be retrieved when linking statically an externally_built library
with a main program.

Additionally, the ada_bind action has been improved so that linker options are
now added directly by the binder, rather than during the post-bind action.