Optimize the memory representation of the Path_Name object.
This object being immutable, having unbounded_strings in it adds
unnecessary footprint since each unbounded string is controlled so
copies of one object leads to 5 calls to the string's Adjust
primitive.

This patch changes the object to an unconstrained internal
representation containing only base strings, and refcounted
visible object allowing passing the object by ref (and storing it
easily), while doing so in a way where a single call to Adjust is
needed.

no-issue-check