Workaround historical windows path limit
Historically, win32 api suffered from a 256-ish path length limitation.

This limit is being lifted through various means, and the GNATCOLL.OS
API allows us to benefit from this lift. So this commit is about
maximizing the use of GNATCOLL for file manipulations to allow the
gpr lib and tools to not be limited by this historical limitation.

Fixes eng/gpr/gpr-issues#620