Reorder slightly the link command line and filter static/shared-libgcc
The linker's required switch attribute is expected to appear after
the linker options retrieved from the views and SALs. Also this may
result in contradicting -static-libgcc and -shared-libgcc options
on the command line, only the last switch is to be used.

This MR fixes issues on x86-windows when handling exceptions on
mixed Ada/C++ applications.

Fixes eng/gpr/gpr-issues#622