Refactor the support for standalone libraries
This change moves the automatic extension of standalone library
interfaces from the link action to the bind action.

This allows simplifying the link action that is already quite
complex, but also simplifies the update of the dependencies from
the ALI files produced by the compile actions.

This also better checks compilation unit traceability by clearly
separating the dependencies coming from the simple Ada
parser from the more accurate dependencies coming from the ALI
files.

Finally this also properly implements the complete support for the
Roots attribute, both in case of libraries or regular projects with
mains.