Fix issue with tempered cursors in case of separates of separates
In the view_tables, iteration on the original Data.Sources container
may lead to issues with tempered cursors when those sources become
hidden due to a renaming of the enclosed unit name.

This is now fixed by using a copy of this container for the iteration.

For eng/gpr/gpr-issues#646