Fix an issue in the makefile parser
Previously in gpr1build when there was a missmatch between the object name and
the object name embedded in the dependency file the recompilation was re-triggered.
In gpr2build we were simply failing the process.
If the Deps are empty this is not a blocking issue for the rest of the build, this
compile action should be done all the time since its signature will never be complete.

Issue: eng/gpr/gpr-issues#646