Command_Line: strengthen handling of response files
This chage modifies the way argument kind are handled by changing
the base structure to match the other similar structures (Cmd_Line
and In_Signature). This allows proper filter out of arguments.

This also expects calls to Recompute_For_Response_File to be called
after the response files are actually created, since the arguments
in the command line are dismissed at this point.