Check that the drivers are correctly defined during tree loading
When loading a tree in autoconf mode, warnings are raised when no
compiler is found for a specific language, target, and runtime. However,
no such warnings are raised when a configuration is provided explicitly.
Although the user is responsible for providing a correct configuration,
checks of this type can be useful to help the user debug their
configuration.

This new warning is disabled for the GPR tools.

Fixes: eng/gpr/gpr-issues#658