GPRDriver: Initial version.
This tools make it possible to call tools either based on GPR1 or GPR2.

This is controlled by the environment variable GNAT_GPR_ENGINE:

- GPR1 if set to "1", "legacy" or "", or if GNAT_GPR_ENGINE undefined.
- GPR2 if set to "2" or "new".

The command line option --gpr=<n> can also be used to select the engine
version to use. This option override the environment variable.

eng/gpr/gpr-issues#686