Do not create the Link_Options_Insert action for shared standalone libraries
This action will insert the options obtained from the binder in a new
linker text section. This is used to retrieve the options required to
link with static standalone externally built libraries, as the binder generated
source containing the options is not available anymore.

This is not required to link with standalone shared libraries. Moreover, this
section insertion is not handled correctly by objcopy for DLLs.

Issue: eng/gpr/gpr-issues#697