Fix the Library_Options attribute processing
This attribute may be used to specify additional switches (“last switches”)
when linking a shared library or a static standalone library. In the case of a
simple static library, the values for this attribute are restricted to paths
to object files. Those paths may be absolute or relative to the object
directory.

Moreover, the attribute processing has been moved to the link actions
instead of the DAG population, to avoid spreading the information
between the two packages.