Actions.Compile.Ada: update Self in the tree in Post_Command
This is needed in particular if the seen output changes after
execution, in case the command was inherited by an extended
view.

This fixes an issue where an action that is being updated can use an
old version of Self to update the DAG, thus bringing obsolete data for
it.

This happens in particular if the compilation of a unit depends on
a child unit (that will depend on the initial unit).

Fixes eng/gpr/gpr-issues#687