Allow loading a signature without checking checksums
Until now, loading a signature would always invalidate it if the checksums of
the current artifacts did not match those stored in the signature. However,
gprls needs to determine the status of all artifacts, so a complete signature
is always required, regardless of checksum consistency.

This commit also replaces the Load_Mode parameter with a more explicit name:
Check_Checksums.

Issue: eng/gpr/gpr-issues#566