Better fix for supporting src-subdirs bindings in SAL.
This changes the way src-subdirs are exported for standalone libraries
by removing the special-cases to allow importing units that are not
part of the interface, but instead automatically adds all units from
the src-subdirs in the interface.

This allows a cleaner bind and check mechanism, while still allowing
instrumentation to be shared between libraries and executables.