Raise a warning when a skipped action has an invalid signature
An action with an invalid signature is not always executed. It can be
skipped or deactivated in the following cases:
- Certain switches (e.g., -b, -c, -l) can prevent specific actions from being
executed.
- The action's command line is empty.
- The action itself determines that it should be skipped.

Before this patch, actions that depended on such skipped actions were
silently not executed.

With this change , a warning is now raised if a skipped or deactivated action
has at least one non-skipped, active descendant.

Fixes: eng/gpr/gpr-issues#709