Provide more information to the user when a project file is not found
The paths used for project-file searching are now displayed to the user as a
categorized list when a missing relative path is passed to the Load subprogram.

To achieve this, the origin of the paths needed to be stored, which is now
done in the Default field of the All_Search_Paths record.

Additional paths are added later in the Recursive_Load subprogram and are
also displayed in their respective categories.

We no longer depend on Search_Paths to ensure that the paths used for searching
match those displayed; previously, these could easily diverge if the
implementation of Search_Paths changed without the display code being updated.

Issues:
- eng/gpr/gpr-issues#682
- eng/gpr/gpr-issues#707
- eng/das/cov/gnatcoverage#406