Updates to gpr2 sources to accommodate limited predefined reference types
Various reference types declared in gpr2 sources must be changed to be
limited types due to their full types having components of reference types
coming from the predefined container packages, which are now declared as
limited types (as required by AI22-0082). Also, the associated reference
functions's bodies are revised to avoid copying limited references within
return statements.

Note that these changes are being made in advance of merging the changes
to the reference types in the predefined container generics in the GNAT
library, but are believed to be compatible with the nonlimited reference
types in the current library.

Issue: eng/toolchain/gnat#1098