Provide a more precise host platform for testing purposes
`gprconfig --help` includes the following message:

```
 --target=TARGET        Select specified target or "all" for any target
   (XXX by default)
```

The obtained value XXX is more precise than the one obtained with the provided
env. For instance, a x86-windows host platform could be detected as
x86_64-windows in the env.host.platform, while it is correctly detected
as x86-windows with gprconfig.

Issue: eng/gpr/gpr-issues#712