Add ALI files to the library linking's signature
When an Ada library depends on another project, changes in that dependency may
affect the library’s output. However, this is not always the case. For example,
when a new API is added to the dependency but is not used by the Ada library,
the object files (.o) embedded in the library remain unchanged, but the ALI
files differ because the dependency has changed.

To be used by the binder, a library must provide up-to-date ALI files in its
library directory. The ALI files are copied from the object directory to the
library directory during the link action’s Pre_Command subprogram. However,
this subprogram is not executed if the action is skipped.

To resolve this issue, all ALI files of a library are now included in its
linking signature.

Fixes: eng/gpr/gpr-issues#742