Remove the C language from the language list for the relocatable library type
The relocatable version of libgpr2 is standalone and does not include
gpr_parser-gdb.c, the only C source file in libgpr2.

By removing C from the language list, we eliminate the warning
"there are no sources of language 'C' in this project" when working with the
relocatable libgpr2.

gpr_parser-gdb.c is used only to support debugging of the Langkit tree.
It is acceptable to provide this file only for the static and static-pic
library variants.

Fixes: eng/gpr/gpr-issues#763