Add compile actions of overridden source from the runtime
When runtime sources are overridden by sources of the project, add a compile action,
the resulting output is added to the final link before linking the runtime and the
ALI is added to the bind action of the main to discover other dependencies.

Minor: Relax the rules of having library dir not equals object dir for
externally built projects and aggregate libraries since their object dir
won't be used.

Issue: eng/gpr/gpr-issues#673